+++ internal/cli/atlas/maintenance/maintenance_test.go
@@ -12,9 +12,19 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package maintenancewindows
+package maintenance
 
-const (
-	maintenanceWindows        = "Manage Atlas maintenance windows."
-	maintenanceWindowsArchive = "Update the maintenance window."
+import (
+	"testing"
+
+	"github.com/mongodb/mongocli/internal/cli"
 )
+
+func TestBuilder(t *testing.T) {
+	cli.CmdValidator(
+		t,
+		Builder(),
+		1,
+		[]string{},
+	)
+}
